Refactoring the files
