--- src/main/java/home/riderly/Controllers/Admin/AdminController.java
@@ -1,4 +1,0 @@
-package home.riderly.Controllers.Admin;
-
-public class AdminController {
-}
