+++ src/main/java/home/riderly/Controllers/User/UserController.java
@@ -1,4 +1,4 @@
-package home.riderly.Controllers.Client;
+package home.riderly.Controllers.User;
 
 import home.riderly.Models.Model;
 import javafx.fxml.Initializable;
@@ -7,7 +7,7 @@ import javafx.scene.layout.BorderPane;
 import java.net.URL;
 import java.util.ResourceBundle;
 
-public class ClientController implements Initializable {
+public class UserController implements Initializable {
     public BorderPane clientParent;
     @Override
     public void initialize(URL url, ResourceBundle resourceBundle) {
