+++ src/main/resources/Fxml/User/Biciclete.fxml
@@ -8,7 +8,7 @@
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.text.Text?>
 
-<AnchorPane prefHeight="750.0" prefWidth="850.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="home.riderly.Controllers.Client.BicicleteController">
+<AnchorPane prefHeight="750.0" prefWidth="850.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="home.riderly.Controllers.User.BicicleteController">
    <Text layoutX="49.0" layoutY="44.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Alege bicicleta:" AnchorPane.leftAnchor="30.0" AnchorPane.topAnchor="30.0" />
    <ImageView fx:id="bicImg1" fitHeight="150.0" fitWidth="225.0" layoutX="90.0" layoutY="127.0" pickOnBounds="true" preserveRatio="true" AnchorPane.leftAnchor="90.0" AnchorPane.topAnchor="126.0">
       <Image url="@../../Images/bic1.jpg" />
