+++ src/main/resources/Fxml/User/Inchiriaza.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.text.Text?>
 
 <AnchorPane prefHeight="126.0" prefWidth="599.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1"
-            fx:controller="home.riderly.Controllers.Client.InchiriazaController">
+            fx:controller="home.riderly.Controllers.User.InchiriazaController">
    <Text layoutX="241.0" layoutY="27.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Ai inchiriat un vehicul!"
          AnchorPane.topAnchor="14.0"/>
    <Text layoutX="127.0" layoutY="50.0" strokeType="OUTSIDE" strokeWidth="0.0"
