+++ src/main/resources/Fxml/User/Istoric.fxml
@@ -6,7 +6,7 @@
 
 
 <AnchorPane prefHeight="750.0" prefWidth="850.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1"
-            fx:controller="home.riderly.Controllers.Client.IstoricController">
+            fx:controller="home.riderly.Controllers.User.IstoricController">
     <Text layoutX="385.0" layoutY="19.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Istoric inchirieri"
           AnchorPane.leftAnchor="385.0" AnchorPane.topAnchor="6.05078125"/>
     <ListView fx:id="istoricListView" layoutX="10.0" layoutY="59.0" prefHeight="640.0" prefWidth="830.0"
