+++ src/main/resources/Fxml/User/User.fxml
@@ -9,7 +9,7 @@
 <BorderPane fx:id="clientParent" xmlns="http://javafx.com/javafx"
             xmlns:fx="http://javafx.com/fxml">
     <left>
-            <fx:include source="ClientMenu.fxml"/>
+            <fx:include source="UserMenu.fxml"/>
     </left>
 
     <center>
