+++ src/main/resources/Fxml/User/UserMenu.fxml
@@ -8,7 +8,7 @@
 <?import javafx.scene.shape.Line?>
 <?import javafx.scene.text.Text?>
 
-<VBox maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="750.0" prefWidth="180.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="home.riderly.Controllers.Client.ClientMenuController">
+<VBox maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="750.0" prefWidth="180.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="home.riderly.Controllers.User.UserMenuController">
     <AnchorPane prefHeight="730.0" prefWidth="160.0">
         <VBox prefHeight="400.0" prefWidth="160.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="100.0">
             <Button fx:id="bicBtn" mnemonicParsing="false" text="Biciclete">
@@ -33,7 +33,7 @@
                 </graphic>
             </Button>
         </VBox>
-        <VBox prefHeight="80.0" prefWidth="160.0" styleClass="title_container" stylesheets="@../../Styles/ClientMenu.css" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
+        <VBox prefHeight="80.0" prefWidth="160.0" styleClass="title_container" stylesheets="@../../Styles/UserMenu.css" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
             <FontAwesomeIconView glyphName="BICYCLE" size="30" />
             <Text strokeType="OUTSIDE" strokeWidth="0.0" text="riderly" />
         </VBox>
